requirements.txt
streamlit==1.6.0
requests==2.27.1

pandas~=1.4.2
numpy~=1.22.3
scikit-learn~=1.0.2
Pillow~=9.1.0
matplotlib~=3.5.2
keras~=2.9.0
tensorflow~=2.9.0
seaborn~=0.11.2
opencv-python